pyproject.toml
[tool.poetry]
name = "task-manager"
version = "0.1.0"
description = ""
authors = ["Aleksey525 <alex_tolchin@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = "^0.24.0"
sqlalchemy = "^2.0.43"
psycopg2-binary = "^2.9.9"
alembic = "^1.16.4"
python-dotenv = "^1.0.0"
asyncpg = "^0.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
